det går nu att sätta ut små röda cirklar
