+++ simulation/event_handler.py
@@ -15,3 +15,6 @@ def handle_events(events: list[Event]):
         if event.key == pygame.K_ESCAPE:
             pygame.quit()
             sys.exit()
+
+    if pygame.mouse.get_pressed()[0] == True:  # Kollar ifall mouse1 är nedtryckt
+        return "mouse 1"
